Moved useProductDetail to its own file
