+++ src/pages/ProductDetail.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { useParams } from "react-router-dom";
 import {
   Container,
@@ -10,33 +10,12 @@ import {
   Popup,
   Placeholder,
 } from "semantic-ui-react";
-import { useSelector, useDispatch } from "react-redux";
-import { selectProductDetail, selectStatus } from "store/products/selectors";
-import { fetchProduct } from "store/products/slice";
-import { STATUS } from "store/products/constants";
 import PriceTag from "components/PriceTag";
-
-function useProductDetail(pid) {
-  const product = useSelector((state) => selectProductDetail(state, pid));
-  const dispatch = useDispatch();
-  const status = useSelector(selectStatus);
-  const isEmpty = status === STATUS.INITIAL && !product;
-  useEffect(
-    function loadProductDetails() {
-      if (isEmpty) {
-        dispatch(fetchProduct(pid));
-      }
-    },
-    [isEmpty, dispatch, pid]
-  );
-
-  return product || {};
-}
+import useProductDetail from "hooks/useProductDetail";
 
 export default function ProductDetail() {
   const { pid } = useParams();
   const product = useProductDetail(pid);
-  console.log("ProductDetail", product);
   return (
     <>
       <Header as="header" textAlign="center">
